Added Battery and Medkit
Added battery and medkit.

Flashlight now has a timer set on how long it lasts, and warns players at percentage levels 75, 50, and 25. At 0 it turns off. Batteres return 50% of battery.

Medkits heal half of player HP when used.

Both items can be found and added to inventory.

Set Common Event 6 to turn off and on the flashlight, but couldn't get it to work with the common event buttons plugin.