Added Stamina System
Added a system so when you press shift it allows you to dash.

Only works if you press, not hold down. So it's not the best system but its the only way I could get it setup.

Can dash for 1 second, takes 3 seconds to fully regen. Good for closing short distances.