Movie Working and Fixed Overworld Zombies
Got movie to play, but unsure if it still has sound.

Movie has its own map, otherwise it freezs, Set player location to that map when game starts, and set event to transfer player to start position after movie plays.

Fixed overworld zombies so that correct movement works for Zombie 6, and that the damage cooldown is setup for them